address comments and move to service module
