+++ project/CelebornBuild.scala
@@ -498,7 +498,6 @@ object CelebornCommon {
 
   lazy val common = Project("celeborn-common", file("common"))
     .dependsOn(CelebornSpi.spi)
-    .dependsOn(CelebornOpenApi.openApiClient % "test->test;compile->compile")
     .settings (
       commonSettings,
       protoSettings,
