+++ service/src/test/scala/org/apache/celeborn/server/common/container/ContainerInfoSuite.scala
@@ -15,31 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.container
+package org.apache.celeborn.server.common.container
 
-import scala.collection.JavaConverters._
-
+import org.apache.celeborn.CelebornFunSuite
 import org.apache.celeborn.common.CelebornConf
-import org.apache.celeborn.common.internal.Logging
 import org.apache.celeborn.common.util.Utils
-import org.apache.celeborn.rest.v1.model.ContainerInfo
 
-abstract class ContainerInfoProvider {
+class ContainerInfoSuite extends CelebornFunSuite {
 
-  def getContainerInfo(): ContainerInfo
-
-}
-
-object ContainerInfoProvider extends Logging {
-
-  val DEFAULT_CONTAINER_NAME = "default_container_name"
-  val DEFAULT_CONTAINER_DATA_CENTER = "default_container_data_center"
-  val DEFAULT_CONTAINER_AVAILABILITY_ZONE = "default_container_availability_zone"
-  val DEFAULT_CONTAINER_CLUSTER = "default_container_cluster"
-  val DEFAULT_CONTAINER_TAGS = List.empty[String].asJava
-
-  def instantiate(conf: CelebornConf): ContainerInfoProvider = {
-    Utils.instantiate(conf.containerInfoProviderClass)
+  test("test DefaultContainerInfoProvider") {
+    val conf = new CelebornConf
+    val defaultContainerInfo = ContainerInfoProvider.instantiate(conf).getContainerInfo()
+    assert(defaultContainerInfo.getContainerHostName == Utils.getHostName(false))
+    assert(defaultContainerInfo.getContainerAddress == Utils.getHostName(true))
+    assert(defaultContainerInfo.getContainerUser == sys.env("user.name"))
   }
-
 }
